fix(admin): update imports after session module split
